Move WSCRS to utils
